requirements.txt
markdown
playwright
python-dotenv
schedule
streamlit
requests